+++ src/vue/modules/movements-history/wrappers/operation-details/payment.js
@@ -10,17 +10,50 @@ export class PaymentOp {
     this.destinationCalculatedPercentFee = record.destinationFee
       .calculatedPercent
     this.sourcePayForDestination = record.sourcePayForDestination
-    this.subject = record.subject
+    this.subject = this._getSubject(record)
     this.reference = record.reference
 
-    this.fromAccountId = record.accountFrom.id
-    this.toAccountId = record.accountTo.id
+    this.fromAccountId = this._getFromAccountId(record)
+    this.toAccountId = this._getToAccountId(record)
     this.fromBalanceId = record.balanceFrom.id
     this.toBalanceId = record.balanceTo.id
     this.assetCode = record.asset.id
   }
 
   isIncoming (accountId) {
-    return this.toAccountId === accountId
+    return this.fromAccountId !== accountId
+  }
+
+  _getSubject (record) {
+    try {
+      return JSON.parse(record.subject).subject
+    } catch (e) {
+      return record.subject
+    }
+  }
+  _getFromAccountId (record) {
+    try {
+      let sender = JSON.parse(record.subject).sender
+      if (sender) {
+        return sender
+      } else {
+        return record.accountFrom.id
+      }
+    } catch (e) {
+      return record.accountFrom.id
+    }
+  }
+
+  _getToAccountId (record) {
+    try {
+      let toAccountEmail = JSON.parse(record.subject).email
+      if (toAccountEmail) {
+        return toAccountEmail
+      } else {
+        return record.accountTo.id
+      }
+    } catch (e) {
+      return record.accountTo.id
+    }
   }
 }